chore: remove mock StatefulSet in pkg controllerutil&testutil (#7174)
(cherry picked from commit 8526bf56cb978e1464551302550ed2940ee5edf9)
