+++ controllers/apps/operations/switchover_util.go
@@ -35,6 +35,7 @@ import (
 	appsv1alpha1 "github.com/apecloud/kubeblocks/apis/apps/v1alpha1"
 	"github.com/apecloud/kubeblocks/pkg/constant"
 	"github.com/apecloud/kubeblocks/pkg/controller/component"
+	"github.com/apecloud/kubeblocks/pkg/controller/instanceset"
 	intctrlutil "github.com/apecloud/kubeblocks/pkg/controllerutil"
 )
 
@@ -90,9 +91,9 @@ func needDoSwitchover(ctx context.Context,
 		if err != nil {
 			return false, err
 		}
-		podParent, _ := intctrlutil.ParseParentNameAndOrdinal(pod.Name)
-		siParent, o := intctrlutil.ParseParentNameAndOrdinal(switchover.InstanceName)
-		if podParent != siParent || o < 0 || o >= int32(len(podList.Items)) {
+		podParent, _ := instanceset.ParseParentNameAndOrdinal(pod.Name)
+		siParent, o := instanceset.ParseParentNameAndOrdinal(switchover.InstanceName)
+		if podParent != siParent || o < 0 || o >= len(podList.Items) {
 			return false, errors.New("switchover.InstanceName is invalid")
 		}
 		// If the current instance is already the primary, then no switchover will be performed.
