+++ pkg/controllerutil/pod_utils.go
@@ -21,7 +21,6 @@ package controllerutil
 
 import (
 	"fmt"
-	"regexp"
 	"strconv"
 	"strings"
 	"time"
@@ -43,26 +42,6 @@ const (
 	// PodScheduledFailedTimeout timeout for scheduling failure.
 	PodScheduledFailedTimeout = 30 * time.Second
 )
-
-// statefulPodRegex is a regular expression that extracts the parent StatefulSet and ordinal from the Name of a Pod
-var statefulPodRegex = regexp.MustCompile("(.*)-([0-9]+)$")
-
-// GetParentNameAndOrdinal gets the name of pod's parent StatefulSet and pod's ordinal as extracted from its Name. If
-// the Pod was not created by a StatefulSet, its parent is considered to be empty string, and its ordinal is considered
-// to be -1.
-func GetParentNameAndOrdinal(pod *corev1.Pod) (string, int) {
-	parent := ""
-	ordinal := -1
-	subMatches := statefulPodRegex.FindStringSubmatch(pod.Name)
-	if len(subMatches) < 3 {
-		return parent, ordinal
-	}
-	parent = subMatches[1]
-	if i, err := strconv.ParseInt(subMatches[2], 10, 32); err == nil {
-		ordinal = int(i)
-	}
-	return parent, ordinal
-}
 
 // GetContainerByConfigSpec searches for container using the configmap of config from the pod
 //
@@ -401,11 +380,6 @@ func PodIsReadyWithLabel(pod corev1.Pod) bool {
 	}
 
 	return PodIsReady(&pod)
-}
-
-// PodIsControlledByLatestRevision checks if the pod is controlled by latest controller revision.
-func PodIsControlledByLatestRevision(pod *corev1.Pod, sts *appsv1.StatefulSet) bool {
-	return GetPodRevision(pod) == sts.Status.UpdateRevision && sts.Status.ObservedGeneration == sts.Generation
 }
 
 // GetPodRevision gets the revision of Pod by inspecting the StatefulSetRevisionLabel. If pod has no revision empty
