--- pkg/controllerutil/stateful_set_utils_test.go
@@ -1,55 +1,0 @@
-/*
-Copyright (C) 2022-2024 ApeCloud Co., Ltd
-
-This file is part of KubeBlocks project
-
-This program is free software: you can redistribute it and/or modify
-it under the terms of the GNU Affero General Public License as published by
-the Free Software Foundation, either version 3 of the License, or
-(at your option) any later version.
-
-This program is distributed in the hope that it will be useful
-but WITHOUT ANY WARRANTY; without even the implied warranty of
-MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-GNU Affero General Public License for more details.
-
-You should have received a copy of the GNU Affero General Public License
-along with this program.  If not, see <http://www.gnu.org/licenses/>.
-*/
-
-package controllerutil
-
-import (
-	"testing"
-
-	testk8s "github.com/apecloud/kubeblocks/pkg/testutil/k8s"
-)
-
-func TestGetParentNameAndOrdinal(t *testing.T) {
-	set := testk8s.NewFakeStatefulSet("foo", 3)
-	pod := testk8s.NewFakeStatefulSetPod(set, 1)
-	if parent, ordinal := GetParentNameAndOrdinal(pod); parent != set.Name {
-		t.Errorf("Extracted the wrong parent name expected %s found %s", set.Name, parent)
-	} else if ordinal != 1 {
-		t.Errorf("Extracted the wrong ordinal expected %d found %d", 1, ordinal)
-	}
-	pod.Name = "1-bar"
-	if parent, ordinal := GetParentNameAndOrdinal(pod); parent != "" {
-		t.Error("Expected empty string for non-member Pod parent")
-	} else if ordinal != -1 {
-		t.Error("Expected -1 for non member Pod ordinal")
-	}
-}
-
-func TestIsMemberOf(t *testing.T) {
-	set := testk8s.NewFakeStatefulSet("foo", 3)
-	set2 := testk8s.NewFakeStatefulSet("bar", 3)
-	set2.Name = "foo2"
-	pod := testk8s.NewFakeStatefulSetPod(set, 1)
-	if !isMemberOf(set, pod) {
-		t.Error("isMemberOf returned false negative")
-	}
-	if isMemberOf(set2, pod) {
-		t.Error("isMemberOf returned false positive")
-	}
-}
